+++ persistent/_compat.py
@@ -13,8 +13,11 @@
 ##############################################################################
 
 import sys
+import os
 
-if sys.version_info[0] > 2: #pragma NO COVER
+PURE_PYTHON = os.environ.get('PURE_PYTHON')
+
+if sys.version_info[0] > 2: # pragma: no cover
     import copyreg as copy_reg
     from collections import UserDict as IterableUserDict
     from collections import UserList
@@ -36,7 +39,7 @@ if sys.version_info[0] > 2: #pragma NO COVER
     PYTHON3 = True
     PYTHON2 = False
 
-else: #pragma NO COVER
+else: # pragma: no cover
     import copy_reg
     from UserDict import IterableUserDict
     from UserList import UserList